+++ pallets/ddc-clusters/src/benchmarking.rs
@@ -91,7 +91,15 @@ benchmarks! {
 								};
 	}: _(RawOrigin::Signed(user.clone()), cluster_id, new_cluster_params)
 	verify {
-		assert_eq!(Clusters::<T>::try_get(cluster_id).unwrap().props, ClusterProps { node_provider_auth_contract: Some(user_2) });
+		assert_eq!(
+			Clusters::<T>::try_get(cluster_id).unwrap().props,
+			ClusterProps {
+				node_provider_auth_contract: Some(user_2),
+				erasure_coding_required: 4,
+				erasure_coding_total: 6,
+				replication_total: 3
+			}
+		);
 	}
 
 	set_cluster_gov_params {