+++ pallets/ddc-clusters/src/mock.rs
@@ -284,7 +284,7 @@ impl ExtBuilder {
 				node_provider_auth_contract: Some(AccountId::from([0; 32])),
 				erasure_coding_required: 4,
 				erasure_coding_total: 6,
-				replication_total: 3
+				replication_total: 3,
 			},
 		) {
 			let _ = pallet_ddc_clusters::GenesisConfig::<Test> {