+++ pallets/ddc-clusters/src/testing_utils.rs
@@ -24,7 +24,7 @@ where
 		node_provider_auth_contract: Some(user.clone()),
 		erasure_coding_required: 4,
 		erasure_coding_total: 6,
-		replication_total: 3
+		replication_total: 3,
 	};
 	let cluster_gov_params: ClusterGovParams<BalanceOf<T>, BlockNumberFor<T>> = ClusterGovParams {
 		treasury_share: Perquintill::default(),
@@ -61,7 +61,7 @@ where
 		node_provider_auth_contract: Some(user.clone()),
 		erasure_coding_required: 4,
 		erasure_coding_total: 6,
-		replication_total: 3
+		replication_total: 3,
 	};
 	let storage_node_params = StorageNodeParams {
 		mode: StorageNodeMode::Storage,
@@ -121,7 +121,7 @@ where
 			node_provider_auth_contract: Some(auth_contract.contract_id),
 			erasure_coding_required: 4,
 			erasure_coding_total: 6,
-			replication_total: 3
+			replication_total: 3,
 		};
 	// Register auth contract
 	let _ = DdcClusters::<T>::set_cluster_params(