+++ pallets/ddc-clusters/src/tests.rs
@@ -61,10 +61,10 @@ fn create_cluster_works() {
 			cluster_manager_id.clone(),
 			cluster_reserve_id.clone(),
 			ClusterParams {
-					node_provider_auth_contract: Some(auth_contract.clone()),
-					erasure_coding_required: 4,
-					erasure_coding_total: 6,
-					replication_total: 3
+				node_provider_auth_contract: Some(auth_contract.clone()),
+				erasure_coding_required: 4,
+				erasure_coding_total: 6,
+				replication_total: 3
 			},
 			cluster_gov_params.clone()
 		));
@@ -169,10 +169,10 @@ fn add_and_delete_node_works() {
 			cluster_manager_id.clone(),
 			cluster_reserve_id.clone(),
 			ClusterParams {
-					node_provider_auth_contract: Some(cluster_manager_id.clone()),
-					erasure_coding_required: 4,
-					erasure_coding_total: 6,
-					replication_total: 3
+				node_provider_auth_contract: Some(cluster_manager_id.clone()),
+				erasure_coding_required: 4,
+				erasure_coding_total: 6,
+				replication_total: 3
 			},
 			ClusterGovParams {
 				treasury_share: Perquintill::from_float(0.05),
@@ -240,10 +240,10 @@ fn add_and_delete_node_works() {
 			RuntimeOrigin::signed(cluster_manager_id.clone()),
 			cluster_id,
 			ClusterParams {
-					node_provider_auth_contract: Some(contract_id),
-					erasure_coding_required: 4,
-					erasure_coding_total: 6,
-					replication_total: 3
+				node_provider_auth_contract: Some(contract_id),
+				erasure_coding_required: 4,
+				erasure_coding_total: 6,
+				replication_total: 3
 			},
 		));
 
@@ -404,10 +404,10 @@ fn set_cluster_params_works() {
 			cluster_manager_id.clone(),
 			cluster_reserve_id.clone(),
 			ClusterParams {
-					node_provider_auth_contract: Some(auth_contract_1),
-					erasure_coding_required: 4,
-					erasure_coding_total: 6,
-					replication_total: 3
+				node_provider_auth_contract: Some(auth_contract_1),
+				erasure_coding_required: 4,
+				erasure_coding_total: 6,
+				replication_total: 3
 			},
 			ClusterGovParams {
 				treasury_share: Perquintill::from_float(0.05),
@@ -437,19 +437,64 @@ fn set_cluster_params_works() {
 			Error::<Test>::OnlyClusterManager
 		);
 
+		assert_noop!(
+			DdcClusters::set_cluster_params(
+				RuntimeOrigin::signed(cluster_manager_id.clone()),
+				cluster_id,
+				ClusterParams {
+					node_provider_auth_contract: Some(auth_contract_2.clone()),
+					erasure_coding_required: 1,
+					erasure_coding_total: 6,
+					replication_total: 3
+				},
+			),
+			Error::<Test>::ErasureCodingRequiredDidNotMeetMinimum
+		);
+
+		assert_noop!(
+			DdcClusters::set_cluster_params(
+				RuntimeOrigin::signed(cluster_manager_id.clone()),
+				cluster_id,
+				ClusterParams {
+					node_provider_auth_contract: Some(auth_contract_2.clone()),
+					erasure_coding_required: 4,
+					erasure_coding_total: 1,
+					replication_total: 3
+				},
+			),
+			Error::<Test>::ErasureCodingTotalNotMeetMinimum
+		);
+
+		assert_noop!(
+			DdcClusters::set_cluster_params(
+				RuntimeOrigin::signed(cluster_manager_id.clone()),
+				cluster_id,
+				ClusterParams {
+					node_provider_auth_contract: Some(auth_contract_2.clone()),
+					erasure_coding_required: 4,
+					erasure_coding_total: 6,
+					replication_total: 1
+				},
+			),
+			Error::<Test>::ReplicationTotalDidNotMeetMinimum
+		);
+
 		assert_ok!(DdcClusters::set_cluster_params(
 			RuntimeOrigin::signed(cluster_manager_id),
 			cluster_id,
 			ClusterParams {
-					node_provider_auth_contract: Some(auth_contract_2.clone()),
-					erasure_coding_required: 4,
-					erasure_coding_total: 6,
-					replication_total: 3
+				node_provider_auth_contract: Some(auth_contract_2.clone()),
+				erasure_coding_required: 4,
+				erasure_coding_total: 6,
+				replication_total: 3
 			},
 		));
 
 		let updated_cluster = DdcClusters::clusters(cluster_id).unwrap();
 		assert_eq!(updated_cluster.props.node_provider_auth_contract, Some(auth_contract_2));
+		assert_eq!(updated_cluster.props.erasure_coding_required, 4);
+		assert_eq!(updated_cluster.props.erasure_coding_total, 6);
+		assert_eq!(updated_cluster.props.replication_total, 3);
 
 		// Checking that event was emitted
 		assert_eq!(System::events().len(), 2);
@@ -496,10 +541,10 @@ fn set_cluster_gov_params_works() {
 			cluster_manager_id.clone(),
 			cluster_reserve_id,
 			ClusterParams {
-					node_provider_auth_contract: Some(auth_contract),
-					erasure_coding_required: 4,
-					erasure_coding_total: 6,
-					replication_total: 3
+				node_provider_auth_contract: Some(auth_contract),
+				erasure_coding_required: 4,
+				erasure_coding_total: 6,
+				replication_total: 3
 			},
 			cluster_gov_params.clone()
 		));
@@ -607,10 +652,10 @@ fn cluster_visitor_works() {
 			cluster_manager_id,
 			cluster_reserve_id.clone(),
 			ClusterParams {
-					node_provider_auth_contract: Some(auth_contract),
-					erasure_coding_required: 4,
-					erasure_coding_total: 6,
-					replication_total: 3
+				node_provider_auth_contract: Some(auth_contract),
+				erasure_coding_required: 4,
+				erasure_coding_total: 6,
+				replication_total: 3
 			},
 			cluster_gov_params
 		));
@@ -719,10 +764,10 @@ fn cluster_creator_works() {
 			cluster_manager_id,
 			cluster_reserve_id,
 			ClusterParams {
-					node_provider_auth_contract: Some(auth_contract),
-					erasure_coding_required: 4,
-					erasure_coding_total: 6,
-					replication_total: 3
+				node_provider_auth_contract: Some(auth_contract),
+				erasure_coding_required: 4,
+				erasure_coding_total: 6,
+				replication_total: 3
 			},
 			cluster_gov_params
 		));