+++ pallets/ddc-staking/src/testing_utils.rs
@@ -115,7 +115,7 @@ pub fn create_stash_controller_node_with_balance<T: Config>(
 		node_provider_auth_contract: Some(stash.clone()),
 		erasure_coding_required: 4,
 		erasure_coding_total: 6,
-		replication_total: 3
+		replication_total: 3,
 	};
 	let cluster_gov_params: ClusterGovParams<BalanceOf<T>, BlockNumberFor<T>> = ClusterGovParams {
 		treasury_share: Perquintill::default(),