+++ CHANGELOG.md
@@ -10,7 +10,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [D] Changes is `Cere Dev` Runtime
 
 ## [VNext]
-- [C,D] `pallet-ddc-clusters`: Added Erasure coding and Replication in cluster params
 
 ## [5.3.0]
 
@@ -18,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [C,D] Updated Substrate to polkadot-v1.1.0
 - [C,D] Introduction of the OpenGov
+- [C,D] `pallet-ddc-clusters`: Added Erasure coding and Replication in cluster params
 
 ## [5.2.0]